backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
dnspython==2.4.2
certifi
PyMuPDF==1.23.8
python-multipart==0.0.6
